+++ intelmq/lib/upgrades.py
@@ -803,6 +803,7 @@ def v310_feed_changes(configuration, harmonization, dry_run, **kwargs):
     found_abusech_feodotracker_csv = []
     found_abusech_feodotracker_browse = []
     found_viriback = []
+    found_netlab_mirai_scanner = []
     messages = []
     for bot_id, bot in configuration.items():
         if bot_id == 'global':
@@ -821,6 +822,8 @@ def v310_feed_changes(configuration, harmonization, dry_run, **kwargs):
                 found_abusech_feodotracker_csv.append(bot_id)
             if http_url == "http://tracker.viriback.com/":
                 found_viriback.append(bot_id)
+            if http_url.startswith("http://data.netlab.360.com/feeds/mirai-scanner/scanner.list"):
+                found_netlab_mirai_scanner.append(bot_id)
         if bot["module"] == "intelmq.bots.parsers.autoshun.parser":
             found_autoshun.append(bot_id)
         if bot["module"] == "intelmq.bots.parsers.dshield.parser_domain":
@@ -862,6 +865,10 @@ def v310_feed_changes(configuration, harmonization, dry_run, **kwargs):
         messages.append('The feed "Viriback Unsafe Site" has been replaced. Please see the feed'
                         ' "Viriback C2 Tracker" and'
                         ' adjust your configuration. Affected bots: %s' % ', '.join(found_viriback))
+
+    if found_netlab_mirai_scanner:
+        messages.append('A discontinued feed "Netlab Mirai Scanner" has been found '
+                        'as bot %s.' % ', '.join(sorted(found_netlab_mirai_scanner)))
 
     messages = ' '.join(messages)
     return messages + ' Remove affected bots yourself.' if messages else None, configuration, harmonization